Add files via upload
1.       To understand the level and dimensions of anticipated stigma, self-stigma, enacted stigma (stigma directly experienced) and observed stigma among people diagnosed with TB:
a.      To understand how and the extent to which self-stigma manifests among people diagnosed with TB.
b.      To understand the settings and stages of care in which TB stigma is being experienced and observed by people diagnosed with TB.
2.      To understand the level and dimensions of secondary TB stigma, stigma directly experienced, and stigma observed by family members / primary carers of people diagnosed with TB:
a.      To understand how and the extent to which secondary stigma manifests among family members / primary carers of people diagnosed with TB.
b.      To understand the settings and stages of care in which secondary TB stigma is being experienced and observed by family members / primary carers of people diagnosed with TB.
3.       To understand the level of perceived TB stigma against people diagnosed with TB in communities, and the ways in which stigma is observed by the community:
a.      To understand how and the extent to which stigma against people diagnosed with TB takes place in communities.
b.      To understand the settings and stages of care in which TB stigma against people diagnosed with TB is being observed by community members.
4.      To understand the level and dimensions of perceived TB stigma against people diagnosed with TB in health care settings and stigma against health care workers:
a.      To understand how and the extent to which perceived stigma against people diagnosed with TB manifests in health care facilities.
b.      To understand the settings in which TB stigma is experienced by TB health care workers.
c.       To understand the settings in which TB stigma against health care workers is observed by other TB health care workers.
5.      To understand the extent to which structural stigma (any existing laws/policies, the enforcement of those laws/policies and the corresponding media coverage) could harm or protect people diagnosed with TB.
6.      To support the development of recommendations to address TB stigma in order to reduce peoples’ vulnerability to TB infection, increase peoples’ access to TB services and improve treatment outcomes.